+++ Framework/Core/Src/TimeUtil.cpp
@@ -6,14 +6,14 @@ float Omega::Core::TimeUtil::GetTime()
 	// here initialize the first call because its static and cons, so never change
 	static const auto startTime = std::chrono::high_resolution_clock::now();
 	auto now = std::chrono::high_resolution_clock::now();
-	return std::chrono::duration_cast<std::chrono::milliseconds>(now - startTime).count() / 1000.0f;
+	return static_cast<float>(std::chrono::duration_cast<std::chrono::milliseconds>(now - startTime).count()) / 1000.0f;
 }
 
 float Omega::Core::TimeUtil::GetDeltaTime()
 {
 	static auto lastTime = std::chrono::high_resolution_clock::now();
 	auto currentTime = std::chrono::high_resolution_clock::now();
-	auto deltaTime = std::chrono::duration_cast<std::chrono::milliseconds>(currentTime - lastTime).count() / 1000.0f;
+	auto deltaTime = static_cast<float>(std::chrono::duration_cast<std::chrono::milliseconds>(currentTime - lastTime).count()) / 1000.0f;
 	lastTime = currentTime;
 	return deltaTime;
 }
